Enable use of firego with provided http client connection
Google App Engine has a custom way to obtain a http client.
Firego does not work on Google App Engine with the
http.client that is created in *firego.New().
Following the discussion on #33 this commit implements a breaking
change adding the http client as a parameter to
*firego.New() which takes the Firebase url and an existing
http client as parameters to create a Firebase connection for
use on e.g. Google App Engine. The "old" Firebase connection
can be obtained by passing "nil" as client to the constructor.