Do not partition the code into multiple file if unnecessary
Go allows you to put multiple types, etc to keep the source code
readable without having to jump to other files. firego is very hard
to read and requires switches to multiple files as you are reading
the source. I am merging a couple of things together to help the
readability.